Remove Download helper
